started working on distance function dvstar. Encountered problem in how cluster container is defined. Will have to be a vector of pointers to vlmc_container.

+++ include/optimize_cmp_kmers/distances/kl_divergence.hpp
@@ -6,7 +6,7 @@ namespace distance {
   
 using vlmc_c = container::VLMC_Container;
 
-float kl(vlmc_c left, vlmc_c right){
+double kl(vlmc_c left, vlmc_c right){
   return 0.0;  
 }
 }
